--- backend/src/tasks/data.js
@@ -1,7 +1,0 @@
-export const tasks = [
-  { id: 1, title: "Drink coffe" },
-  { id: 2, title: "click click from keyboard like I'm working :)" },
-  { id: 3, title: "Изучать Express" },
-  { id: 4, title: "Практиковать React" },
-  { id: 5, title: "sleep like a cat(18 hours per day)" },
-];
